requirements.txt
# Python 依賴套件清單
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
requests==2.31.0
beautifulsoup4==4.12.2
